+++ lib/rspec/core/pending.rb
@@ -1,7 +1,7 @@
 module RSpec
   module Core
     module Pending
-      class PendingDeclaredInExample < StandardError; end
+      class SkipDeclaredInExample < StandardError; end
 
       # If Test::Unit is loaed, we'll use its error as baseclass, so that Test::Unit
       # will report unmet RSpec expectations as failures rather than errors.
@@ -18,12 +18,11 @@ module RSpec
       # @overload pending(message)
       # @overload pending(message, &block)
       #
-      # Stops execution of an example, and reports it as pending. Takes an
-      # optional message and block.
+      # Marks an example as pending. The rest of the example will still be
+      # executed, and if it passes the example will fail to indicate that the
+      # pending can be removed.
       #
       # @param [String] message optional message to add to the summary report.
-      # @param [Block] block optional block. If it fails, the example is
-      #   reported as pending. If it executes cleanly the example fails.
       #
       # @example
       #
@@ -31,28 +30,13 @@ module RSpec
       #       # reported as "Pending: no reason given"
       #       it "is pending with no message" do
       #         pending
-      #         this_does_not_get_executed
+      #         raise "broken" 
       #       end
       #
       #       # reported as "Pending: something else getting finished"
       #       it "is pending with a custom message" do
       #         pending("something else getting finished")
-      #         this_does_not_get_executed
-      #       end
-      #
-      #       # reported as "Pending: something else getting finished"
-      #       it "is pending with a failing block" do
-      #         pending("something else getting finished") do
-      #           raise "this is the failure"
-      #         end
-      #       end
-      #
-      #       # reported as failure, saying we expected the block to fail but
-      #       # it passed.
-      #       it "is pending with a passing block" do
-      #         pending("something else getting finished") do
-      #           true.should be(true)
-      #         end
+      #         raise "broken"
       #       end
       #     end
       #
@@ -74,34 +58,82 @@ module RSpec
       def pending(*args)
         current_example = RSpec.current_example
 
-        return self.class.before(:each) { pending(*args) } unless current_example
+        if current_example
+          Pending.mark_pending! current_example, args.first
+        else
+          self.class.before(:each) { pending(*args) }
+        end
+      end
 
+      # @overload skip()
+      # @overload skip(message)
+      # @overload skip(message, &block)
+      #
+      # Marks an example as pending and skips execution when called without a
+      # block. When called with a block, skips just that block and does not
+      # mark the example as pending. The block form is provided as replacement
+      # for RSpec 2's pending-with-block feature, and is not recommended for
+      # new code. Use simple conditionals instead.
+      #
+      # @param [String] message optional message to add to the summary report.
+      # @block [Block] block optional block to be skipped
+      #
+      # @example
+      #
+      #     describe "an example" do
+      #       # reported as "Pending: no reason given"
+      #       it "is skipped with no message" do
+      #         skip
+      #       end
+      #
+      #       # reported as "Pending: something else getting finished"
+      #       it "is skipped with a custom message" do
+      #         skip "something else getting finished"
+      #       end
+      #
+      #       # Passes
+      #       it "contains a skipped statement" do
+      #         skip do
+      #           fail
+      #         end
+      #       end
+      #     end
+      def skip(*args, &block)
+        return block && block.call if Pending.guarded?(*args)
+        return if block
+
+        current_example = RSpec.current_example
+
+        if current_example
+          Pending.mark_pending! current_example, args
+          current_example.metadata[:skip] = true
+          raise SkipDeclaredInExample
+        else
+          self.class.before(:each) { skip(*args) }
+        end
+      end
+
+      def self.mark_pending!(example, message_or_bool)
+        message = if !message_or_bool || !(String === message_or_bool)
+          NO_REASON_GIVEN
+        else
+          message_or_bool
+        end
+
+        example.metadata[:pending] = true
+        example.metadata[:execution_result][:pending_message] = message
+        example.execution_result[:pending_fixed] = false
+      end
+
+      def self.mark_fixed!(example)
+        example.metadata[:pending] = false
+        example.metadata[:execution_result][:pending_fixed] = true
+      end
+
+      def self.guarded?(*args)
         options = args.last.is_a?(Hash) ? args.pop : {}
-        message = args.first || NO_REASON_GIVEN
 
-        if options[:unless] || (options.has_key?(:if) && !options[:if])
-          return block_given? ? yield : nil
-        end
-
-        current_example.metadata[:pending] = true
-        current_example.metadata[:execution_result][:pending_message] = message
-        current_example.execution_result[:pending_fixed] = false
-        if block_given?
-          begin
-            no_failure = false
-            yield
-            no_failure = true
-            current_example.metadata[:pending] = false
-          rescue Exception => e
-            current_example.execution_result[:exception] = e
-          end
-
-          if no_failure
-            current_example.execution_result[:pending_fixed] = true
-            raise PendingExampleFixedError.new
-          end
-        end
-        raise PendingDeclaredInExample.new(message)
+        options[:unless] || (options.has_key?(:if) && !options[:if])
       end
     end
   end