+++ spec/rspec/core/formatters/base_formatter_spec.rb
@@ -52,7 +52,6 @@ RSpec.describe RSpec::Core::Formatters::BaseFormatter do
       end
 
       it "doesn't hang when file exists" do
-        pending("This issue still exists on JRuby, but should be resolved shortly: https://github.com/rspec/rspec-core/issues/295", :if => RUBY_PLATFORM == 'java')
         exception = double(:Exception, :backtrace => [ "#{__FILE__}:#{__LINE__}"])
 
         example = double(:Example, :file_path => __FILE__)