+++ spec/rspec/core/resources/formatter_specs.rb
@@ -7,17 +7,15 @@ end
 RSpec.describe "pending command with block format" do
   context "with content that would fail" do
     it "is pending" do
-      pending do
-        expect(1).to eq(2)
-      end
+      pending
+      expect(1).to eq(2)
     end
   end
 
   context "with content that would pass" do
     it "fails" do
-      pending do
-        expect(1).to eq(1)
-      end
+      pending
+      expect(1).to eq(1)
     end
   end
 end